Norm : libft, innstall
