+++ src/minishell.h
@@ -6,7 +6,7 @@
 /*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/25 14:17:39 by sham              #+#    #+#             */
-/*   Updated: 2021/12/25 18:39:09 by marvin           ###   ########.fr       */
+/*   Updated: 2021/12/25 20:41:43 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -139,11 +139,15 @@ char	*ft_itoa(long n);
 
 /* install */
 void	fork_cmd(t_list *cmd_list, t_list *env_list);
-int 	handle_dis(t_cmd *cmd);
 int		check_bulit_in(t_cmd *cmd);
 void	parse_cmd(t_cmd *cmd, t_list *env_list);
-void	execve_cmd_bult_in(char *cmd_name, t_cmd *cmd, t_list *env_list, int is_forked);
+void	execve_cmd_bult_in(char *cmd_name, t_cmd *cmd, t_list *env_list, \
+int is_forked);
 void	execve_cmd_normal(char *cmd_name, t_cmd *cmd, t_list *env_list);
+
+/* dis */
+int 	handle_dis(t_cmd *cmd);
+int		ft_d_left_heredoc(char *dst);
 
 /* bulit_in */
 int ft_cd(char *path, t_list *env_list);
